orders/pyproject.toml
[project]
name = "orders"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.12.0",
    "fastapi[standard]>=0.118.3",
    "pydantic-settings>=2.11.0",
    "redis>=6.4.0",
]
